Move non-shared utils to respective files
